requirements.txt
streamlit
langchain
langchain-google-genai
python-docx
PyPDF2
pandas
